requirements.txt
faiss-cpu
streamlit
langchain
langchain_community
pypdf
python-dotenv
pandas
langchain_google_genai
duckduckgo-search
wikipedia
#-e .